refactor: [#59 - ] property 관리 클래스를 global/properties 디렉토리로 리팩토링 - SecurityProperties를 global/properties로 이동 - AuthProperty를 global/properties로 이동 Fixes: #59
