+++ src/test/java/com/komentum/test/MockMvcUtils.java
@@ -5,8 +5,8 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.komentum.auth.AuthProperty;
 import com.komentum.auth.JwtUtils;
+import com.komentum.global.properties.AuthProperty;
 import com.komentum.test.dto.MockMvcMultipartRequestDto;
 import com.komentum.test.dto.MockMvcMultipartRequestDto.MultipartExecutionContext;
 import com.komentum.test.dto.MockMvcRequestDto;
